requirements.txt
Django==1.7.5
GDAL==1.7.3
GnuPGInterface==0.3.2
PyYAML==3.11
Shapely==1.5.6
South==1.0.2
apt-xapian-index==0.44
argparse==1.2.1
chardet==2.0.1
command-not-found==0.2.44
distribute==0.6.24
django-bootstrap3==5.1.1
django-mapit==1.3
language-selector==0.1
numpy==1.6.1
psycopg2==2.6
python-apt==0.8.3ubuntu7
python-debian==0.1.21ubuntu1
ufw==0.31.1-1
virtualenv==1.7.1.2
wsgiref==0.1.2
